4 practice

+++ src/main/java/practice2/Help.java
@@ -4,5 +4,6 @@ import java.rmi.Remote;
 import java.rmi.RemoteException;
 
 public interface Help extends Remote {
-    Answer solveTask(int a, int b, int c) throws RemoteException;
+    Answer solveTask(double a, double b, double c) throws RemoteException;
 }
+
